refactor(layout): split layout transformer into mode-specific modules
layout.ts had grown to handle frame composition, flex conversions, grid
positioning, sizing, dimensions, and grid sort/overlap detection — all in
one file. Pull the mode-specific helpers into their own modules:

- layout/common.ts — SimplifiedLayout type, sizing/align/mode converters,
  gap shorthand
- layout/flex.ts — flex-mode justify / align / gap
- layout/grid.ts — grid utilities (computeGridChildOrder, isPackedGrid,
  gridChildrenOverlap, buildGridChildPositioning)

layout.ts keeps the builder (buildSimplifiedLayout and its inner
buildSimplifiedFrameValues / buildSimplifiedLayoutValues) and re-exports
SimplifiedLayout + computeGridChildOrder so existing import paths still
work. No behavior change.
